move runtime-transaction chain into submodule
